lineboterp/lineboterp_consumer/requirements.txt
line-bot-sdk
flask
gunicorn
mysql-connector-python
Pillow
boto3
pyimgur
schedule